change m_pos to off_t type

+++ include/ResMan.h
@@ -40,7 +40,7 @@ class FileLike
     private:
         unsigned char* m_buf;
         size_t m_size;
-        int m_pos;
+        off_t m_pos;
 };
 
 /*!
